+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RPC v0.4 `starknet_getTransactionReceipt` incorrect execution and finality status names
 - `pathfinder_getTransactionStatus` fails to parse v0.12.1 gateway replies
 
+### Changed
+
+- RPC v0.4.0 support (previously supported v0.4.0-rc3)
+
 ## [0.7.0] - 2023-07-27
 
 ### Added